+++ tests/performance_tests/test_perf_live_execution.py
@@ -30,6 +30,7 @@ from nautilus_trader.live.execution_engine import LiveExecutionEngine
 from nautilus_trader.live.risk_engine import LiveRiskEngine
 from nautilus_trader.model.commands import SubmitOrder
 from nautilus_trader.model.enums import OrderSide
+from nautilus_trader.model.enums import VenueType
 from nautilus_trader.model.identifiers import AccountId
 from nautilus_trader.model.identifiers import ClientId
 from nautilus_trader.model.identifiers import PositionId
@@ -90,6 +91,7 @@ class TestLiveExecutionPerformance(PerformanceHarness):
 
         exec_client = MockExecutionClient(
             client_id=ClientId("BINANCE"),
+            venue_type=VenueType.EXCHANGE,
             account_id=self.account_id,
             engine=self.exec_engine,
             clock=self.clock,